ACS-11592 renamed class to avoid ambiguity with same named class
