+++ live-ingester/src/integration-test/java/org/alfresco/hxi_connector/live_ingester/adapters/messaging/hx_insight/storage/connector/PreSignedUrlRequesterIntegrationTest.java
@@ -94,7 +94,7 @@ import org.alfresco.hxi_connector.common.exception.EndpointClientErrorException;
 import org.alfresco.hxi_connector.common.exception.EndpointServerErrorException;
 import org.alfresco.hxi_connector.common.test.docker.util.DockerContainers;
 import org.alfresco.hxi_connector.common.test.docker.util.DockerTags;
-import org.alfresco.hxi_connector.common.test.util.LoggingUtils;
+import org.alfresco.hxi_connector.common.test.util.LogCaptureUtils;
 import org.alfresco.hxi_connector.live_ingester.IntegrationCamelTestBase;
 import org.alfresco.hxi_connector.live_ingester.adapters.auth.LiveIngesterAuthClient;
 import org.alfresco.hxi_connector.live_ingester.adapters.config.IntegrationProperties;
@@ -279,7 +279,7 @@ class PreSignedUrlRequesterIntegrationTest extends IntegrationCamelTestBase
         // given
         givenThat(post(PRE_SIGNED_URL_PATH)
                 .willReturn(badRequest().withResponseBody(new Body("{\"error\": \"Bad request\"}"))));
-        ListAppender<ILoggingEvent> logEntries = LoggingUtils.createLogsListAppender(PreSignedUrlRequester.class);
+        ListAppender<ILoggingEvent> logEntries = LogCaptureUtils.createLogsListAppender(PreSignedUrlRequester.class);
 
         // when
         catchThrowable(locationRequester::requestStorageLocation);
