+++ live-ingester/src/integration-test/java/org/alfresco/hxi_connector/live_ingester/adapters/messaging/transform/storage/SharedFileStoreClientIntegrationTest.java
@@ -63,7 +63,7 @@ import org.alfresco.hxi_connector.common.config.properties.Application;
 import org.alfresco.hxi_connector.common.exception.EndpointClientErrorException;
 import org.alfresco.hxi_connector.common.exception.EndpointServerErrorException;
 import org.alfresco.hxi_connector.common.test.docker.util.DockerContainers;
-import org.alfresco.hxi_connector.common.test.util.LoggingUtils;
+import org.alfresco.hxi_connector.common.test.util.LogCaptureUtils;
 import org.alfresco.hxi_connector.live_ingester.IntegrationCamelTestBase;
 import org.alfresco.hxi_connector.live_ingester.adapters.config.IntegrationProperties;
 import org.alfresco.hxi_connector.live_ingester.domain.ports.transform_engine.TransformEngineFileStorage;
@@ -159,7 +159,7 @@ class SharedFileStoreClientIntegrationTest extends IntegrationCamelTestBase
         // given
         givenThat(get(SFS_DOWNLOAD_FILE_PATH)
                 .willReturn(badRequest().withResponseBody(new Body("{\"error\": \"Bad request\"}"))));
-        ListAppender<ILoggingEvent> logEntries = LoggingUtils.createLogsListAppender(SharedFileStoreClient.class);
+        ListAppender<ILoggingEvent> logEntries = LogCaptureUtils.createLogsListAppender(SharedFileStoreClient.class);
 
         // when
         catchThrowable(() -> sharedFileStoreClient.downloadFile(FILE_ID));
